requirements.txt
nose==1.2.1
nose-cov==1.6
pyyaml==3.10
